NODE-778: Tests passed
