+++ lang/jvm/src/test/scala/com/wavesplatform/lang/typechecker/CompilerV1Test.scala
@@ -2,7 +2,6 @@ package com.wavesplatform.lang.typechecker
 
 import com.wavesplatform.lang.Common._
 import com.wavesplatform.lang.v1.compiler
-import com.wavesplatform.lang.v1.compiler.CompilerV1.CompilationResult
 import com.wavesplatform.lang.v1.compiler.Terms._
 import com.wavesplatform.lang.v1.compiler.{CompilerContext, CompilerV1}
 import com.wavesplatform.lang.v1.evaluator.ctx.impl.PureContext._
@@ -152,13 +151,13 @@ class CompilerV1Test extends PropSpec with PropertyChecks with Matchers with Scr
       Expressions.LET(0, 0, Expressions.PART.INVALID(0, 1, "can't parse"), Expressions.TRUE(0, 0), Seq.empty),
       Expressions.REF(0, 0, Expressions.PART.VALID(0, 0, "x"))
     ),
-    expectedResult = Left("Typecheck failed: Can't compile an invalid instruction: can't parse in 0-1")
+    expectedResult = Left("Typecheck failed: can't parse in 0-1")
   )
 
   treeTypeTest("Invalid GETTER")(
     ctx = typeCheckerContext,
     expr = Expressions.GETTER(0, 0, Expressions.REF(0, 0, Expressions.PART.VALID(0, 0, "x")), Expressions.PART.INVALID(2, 3, "can't parse")),
-    expectedResult = Left("Typecheck failed: Can't compile an invalid instruction: can't parse in 2-3")
+    expectedResult = Left("Typecheck failed: can't parse in 2-3")
   )
 
   treeTypeTest("Invalid BYTEVECTOR")(
@@ -176,13 +175,13 @@ class CompilerV1Test extends PropSpec with PropertyChecks with Matchers with Scr
   treeTypeTest("Invalid REF")(
     ctx = typeCheckerContext,
     expr = Expressions.REF(0, 0, Expressions.PART.INVALID(0, 0, "can't parse")),
-    expectedResult = Left("Typecheck failed: Can't compile an invalid instruction: can't parse in 0-0")
+    expectedResult = Left("Typecheck failed: can't parse in 0-0")
   )
 
   treeTypeTest("Invalid FUNCTION_CALL")(
     ctx = typeCheckerContext,
     expr = Expressions.FUNCTION_CALL(0, 0, Expressions.PART.INVALID(0, 0, "can't parse"), List.empty),
-    expectedResult = Left("Typecheck failed: Can't compile an invalid instruction: can't parse in 0-0")
+    expectedResult = Left("Typecheck failed: can't parse in 0-0")
   )
 
   treeTypeTest("INVALID")(
@@ -191,7 +190,7 @@ class CompilerV1Test extends PropSpec with PropertyChecks with Matchers with Scr
     expectedResult = Left("Typecheck failed: ### in 0-0")
   )
 
-  private def treeTypeTest(propertyName: String)(expr: Expressions.EXPR, expectedResult: CompilationResult[EXPR], ctx: CompilerContext): Unit =
+  private def treeTypeTest(propertyName: String)(expr: Expressions.EXPR, expectedResult: Either[String, EXPR], ctx: CompilerContext): Unit =
     property(propertyName) {
       compiler.CompilerV1(ctx, expr) shouldBe expectedResult
     }
