+++ lang/shared/src/main/scala/com/wavesplatform/lang/v1/compiler/CompilerV1.scala
@@ -1,26 +1,22 @@
 package com.wavesplatform.lang.v1.compiler
 
-import cats.data._
-import cats.instances.either._
-import cats.instances.list._
-import cats.syntax.all._
+import cats.Show
+import cats.implicits._
 import com.wavesplatform.lang.ExprCompiler
 import com.wavesplatform.lang.ScriptVersion.Versions.V1
 import com.wavesplatform.lang.directives.Directive
 import com.wavesplatform.lang.v1.FunctionHeader
 import com.wavesplatform.lang.v1.FunctionHeader.FunctionHeaderType
+import com.wavesplatform.lang.v1.compiler.CompilationError._
+import com.wavesplatform.lang.v1.compiler.CompilerContext._
 import com.wavesplatform.lang.v1.compiler.Terms._
-import com.wavesplatform.lang.v1.evaluator.ctx._
 import com.wavesplatform.lang.v1.evaluator.ctx.PredefFunction.FunctionTypeSignature
+import com.wavesplatform.lang.v1.evaluator.ctx._
 import com.wavesplatform.lang.v1.evaluator.ctx.impl.PureContext
-import com.wavesplatform.lang.v1.parser.BinaryOperation
 import com.wavesplatform.lang.v1.parser.BinaryOperation._
-import com.wavesplatform.lang.v1.parser.Expressions.BINARY_OP
-import com.wavesplatform.lang.v1.parser.Expressions.PART
-import com.wavesplatform.lang.v1.parser.{Expressions, Parser}
-import monix.eval.Coeval
-
-import scala.util.Try
+import com.wavesplatform.lang.v1.parser.Expressions.{BINARY_OP, MATCH_CASE, PART}
+import com.wavesplatform.lang.v1.parser.{BinaryOperation, Expressions, Parser}
+import com.wavesplatform.lang.v1.task.imports._
 
 class CompilerV1(ctx: CompilerContext) extends ExprCompiler {
   override type V = V1.type
@@ -43,260 +39,241 @@ class CompilerV1(ctx: CompilerContext) extends ExprCompiler {
 
 object CompilerV1 {
 
-  type TypeResolutionError      = String
-  type CompilationResult[T]     = Either[TypeResolutionError, T]
-  private type SetTypeResult[T] = EitherT[Coeval, String, T]
-
-  type ResolvedArgsResult = EitherT[Coeval, String, List[EXPR]]
-
-  private def compile(ctx: CompilerContext, expr: Expressions.EXPR): SetTypeResult[EXPR] = expr match {
-    case x: Expressions.CONST_LONG       => EitherT.pure(CONST_LONG(x.value))
-    case x: Expressions.CONST_BYTEVECTOR => handlePart(x.value)(CONST_BYTEVECTOR)
-    case x: Expressions.CONST_STRING     => handlePart(x.value)(CONST_STRING)
-    case _: Expressions.TRUE             => EitherT.pure(TRUE)
-    case _: Expressions.FALSE            => EitherT.pure(FALSE)
-    case getter: Expressions.GETTER      => compileGetter(ctx, getter)
-    case fc: Expressions.FUNCTION_CALL   => compileFunctionCall(ctx, fc)
-    case block: Expressions.BLOCK        => compileBlock(ctx, block)
-    case ifExpr: Expressions.IF          => compileIf(ctx, ifExpr)
-    case ref: Expressions.REF            => compileRef(ctx, ref)
-    case m: Expressions.MATCH            => compileMatch(ctx, m)
-    case Expressions.BINARY_OP(start, end, a, op, b) =>
-      op match {
-        case AND_OP => compileIf(ctx, Expressions.IF(start, end, a, b, Expressions.FALSE(start, end)))
-        case OR_OP  => compileIf(ctx, Expressions.IF(start, end, a, Expressions.TRUE(start, end), b))
-        case _      => compileFunctionCall(ctx, Expressions.FUNCTION_CALL(start, end, PART.VALID(start, end, opsToFunctions(op)), List(a, b)))
-      }
-    case Expressions.INVALID(start, end, message, _) => EitherT.leftT[Coeval, EXPR](s"$message in $start-$end")
-  }
-
-  private def compileGetter(ctx: CompilerContext, getter: Expressions.GETTER): SetTypeResult[EXPR] =
-    for {
-      field <- EitherT.fromEither[Coeval](getter.field.toEither)
-      r <- compile(ctx, getter.ref)
-        .subflatMap { subExpr =>
-          def resolveFields(tpe: PredefBase): List[(String, TYPE)] = tpe match {
-            case PredefCaseType(_, fields) => fields
-            case UnionType(_, types)       => types.map(n => resolveFields(ctx.predefTypes(n.name)).toSet).reduce(_ intersect _).toList
-          }
-
-          def getField(typeName: String): Either[String, GETTER] = {
-            val refTpe = ctx.predefTypes.get(typeName).map(Right(_)).getOrElse(Left(s"Undefined type: $typeName"))
-            val fieldTpe = refTpe.flatMap { ct =>
-              val fieldTpe = resolveFields(ct).collectFirst { case (fieldName, tpe) if fieldName == field => tpe }
-              fieldTpe.map(Right(_)).getOrElse(Left(s"Undefined field `$field` of variable of type `$typeName`"))
-            }
-            fieldTpe.right.map(tpe => GETTER(expr = subExpr, field = field, tpe = tpe))
-          }
-
-          subExpr.tpe match {
-            case typeRef: CASETYPEREF => getField(typeRef.name)
-            case union: UNION =>
-              val x1 = union.l
-                .map(k => ctx.predefTypes(k.name))
-                .map(predefType => resolveFields(predefType).find(_._1 == field))
-              if (x1.contains(None)) Left(s"Undefined field `$field` on $union")
-              else
-                TypeInferrer.findCommonType(x1.map(_.get._2)) match {
-                  case Some(cT) => Right(GETTER(expr = subExpr, field = field, tpe = cT))
-                  case None     => Left(s"Undefined common type for field `$field` on $union")
-                }
-
-            case x => Left(s"Can't access to '$field' of a primitive type $x")
-          }
+  def compileExpr(expr: Expressions.EXPR): CompileM[Terms.EXPR] = {
+    expr match {
+      case x: Expressions.CONST_LONG                         => (CONST_LONG(x.value): EXPR).pure[CompileM]
+      case x: Expressions.CONST_BYTEVECTOR                   => handlePart(x.value).map(CONST_BYTEVECTOR)
+      case x: Expressions.CONST_STRING                       => handlePart(x.value).map(CONST_STRING)
+      case _: Expressions.TRUE                               => (TRUE: EXPR).pure[CompileM]
+      case _: Expressions.FALSE                              => (FALSE: EXPR).pure[CompileM]
+      case Expressions.GETTER(start, end, ref, field)        => compileGetter(start, end, field, ref)
+      case Expressions.BLOCK(start, end, let, body)          => compileBlock(start, end, let, body)
+      case Expressions.IF(start, end, cond, ifTrue, ifFalse) => compileIf(start, end, cond, ifTrue, ifFalse)
+      case Expressions.REF(start, end, key)                  => compileRef(start, end, key)
+      case Expressions.FUNCTION_CALL(start, end, name, args) => compileFunctionCall(start, end, name, args)
+      case Expressions.MATCH(start, end, ex, cases)          => compileMatch(start, end, ex, cases.toList)
+      case Expressions.INVALID(start, end, message, _)       => raiseError(Generic(start, end, message))
+      case Expressions.BINARY_OP(start, end, a, op, b) =>
+        op match {
+          case AND_OP => compileIf(start, end, a, b, Expressions.FALSE(start, end))
+          case OR_OP  => compileIf(start, end, a, Expressions.TRUE(start, end), b)
+          case _      => compileFunctionCall(start, end, PART.VALID(start, end, BinaryOperation.opsToFunctions(op)), List(a, b))
         }
-    } yield r
-
-  private def compileIf(ctx: CompilerContext, ifExpr: Expressions.IF): SetTypeResult[EXPR] =
-    (compile(ctx, ifExpr.cond), compile(ctx, ifExpr.ifTrue), compile(ctx, ifExpr.ifFalse)).tupled
-      .subflatMap[String, EXPR] {
-        case (resolvedCond: EXPR, resolvedIfTrue, resolvedIfFalse) =>
-          if (resolvedCond.tpe != BOOLEAN)
-            Left(s"IF clause is expected to be BOOLEAN, acutal: ${resolvedCond.tpe}")
-          else {
-            val ifTrueTpe  = resolvedIfTrue.tpe
-            val ifFalseTpe = resolvedIfFalse.tpe
-            TypeInferrer.findCommonType(ifTrueTpe, ifFalseTpe) match {
-              case Some(tpe) =>
-                Right(
-                  IF(
-                    cond = resolvedCond,
-                    ifTrue = resolvedIfTrue,
-                    ifFalse = resolvedIfFalse,
-                    tpe = tpe
-                  ))
-              case None => Left(s"Can't find common type for $ifTrueTpe and $ifFalseTpe")
-            }
-          }
-      }
-
-  private def compileFunctionCall(ctx: CompilerContext, fc: Expressions.FUNCTION_CALL): SetTypeResult[EXPR] = {
-    val Expressions.FUNCTION_CALL(_, _, name, args) = fc
-    for {
-      name <- EitherT.fromEither[Coeval](name.toEither)
-      r <- ctx.functionTypeSignaturesByName(name) match {
-        case Nil                   => EitherT.fromEither[Coeval](Left(s"Function '$name' not found"))
-        case singleOverload :: Nil => resolvedFuncArguments(ctx, args).subflatMap(matchFuncOverload(name, args, _, singleOverload))
-        case many =>
-          resolvedFuncArguments(ctx, args).subflatMap { resolvedArgs =>
-            val matchedSignatures = many
-              .zip(many.map(matchFuncOverload(name, args, resolvedArgs, _)))
-              .collect {
-                case (sig, result) if result.isRight => (sig, result)
-              }
-
-            matchedSignatures match {
-              case Nil                       => Left(s"Can't find a function '$name'(${resolvedArgs.map(_.tpe.typeInfo).mkString(", ")})")
-              case (_, oneFuncResult) :: Nil => oneFuncResult
-              case manyPairs =>
-                val candidates = manyPairs.map { case (sig, _) => s"'$name'(${sig.args.mkString(", ")})" }
-                Left(s"Can't choose an overloaded function. Candidates: ${candidates.mkString("; ")}")
-            }
-          }
-      }
-    } yield r
+    }
   }
 
-  private def compileBlock(ctx: CompilerContext, block: Expressions.BLOCK): SetTypeResult[EXPR] =
+  private def compileIf(start: Int,
+                        end: Int,
+                        condExpr: Expressions.EXPR,
+                        ifTrueExpr: Expressions.EXPR,
+                        ifFalseExpr: Expressions.EXPR): CompileM[Terms.EXPR] = {
+    get[CompilerContext, CompilationError].flatMap(ctx => {
+      (for {
+        cond       <- compileExpr(condExpr).run(ctx).map(_._2)()
+        ifTrue     <- compileExpr(ifTrueExpr).run(ctx).map(_._2)()
+        ifFalse    <- compileExpr(ifFalseExpr).run(ctx).map(_._2)()
+        compiledIf <- mkIf(start, end, cond, ifTrue, ifFalse)
+      } yield compiledIf).toCompileM
+    })
+  }
+
+  def flat(typeDefs: Map[String, PredefBase], tl: List[String]): List[String] =
+    tl.flatMap(typeName =>
+      typeDefs.get(typeName).fold(List(typeName)) {
+        case UnionType(_, types) => types.map(_.name)
+        case simple              => List(simple.name)
+    })
+
+  private def compileMatch(start: Int, end: Int, expr: Expressions.EXPR, cases: List[Expressions.MATCH_CASE]): CompileM[Terms.EXPR] = {
     for {
-      letName <- EitherT.fromEither[Coeval](block.let.name.toEither)
-      r <- (ctx.varDefs.get(letName), ctx.functionDefs.get(letName)) match {
-        case (Some(_), _) => EitherT.leftT[Coeval, EXPR](s"Value '$letName' already defined in the scope")
-        case (_, Some(_)) =>
-          EitherT.leftT[Coeval, EXPR](s"Value '$letName' can't be defined because function with such name is predefined")
-        case (None, None) =>
-          import block.let
-          for {
-            exprTpe  <- compile(ctx, let.value)
-            letTypes <- EitherT.fromEither[Coeval](let.types.map(_.toEither).toList.sequence[CompilationResult, String])
-            _ <- EitherT.cond[Coeval](
-              letTypes.forall(ctx.predefTypes.contains),
+      ctx       <- get[CompilerContext, CompilationError]
+      typedExpr <- compileExpr(expr)
+      exprTypes <- typedExpr.tpe match {
+        case u: UNION => u.pure[CompileM]
+        case _        => raiseError[CompilerContext, CompilationError, UNION](MatchOnlyUnion(start, end))
+      }
+      _ <- cases
+        .flatMap(_.types)
+        .traverse[CompileM, String](handlePart)
+        .map(tl => UNION(flat(ctx.predefTypes, tl).map(CASETYPEREF)))
+        .flatMap(matchedTypes => {
+          Either
+            .cond(
+              (cases.last.types.isEmpty && (exprTypes >= matchedTypes)) || (exprTypes equivalent matchedTypes),
               (),
-              s"Value '$letName' declared as non-existing type $letTypes, while all possible types are ${ctx.predefTypes}"
+              MatchNotExhaustive(start, end, exprTypes.l, matchedTypes.l)
             )
-            desiredUnion = if (let.types.isEmpty) exprTpe.tpe else UNION(letTypes.map(CASETYPEREF))
-            updatedCtx   = ctx.copy(varDefs = ctx.varDefs + (letName -> desiredUnion))
-            inExpr <- compile(updatedCtx, block.body)
-          } yield
-            BLOCK(
-              let = LET(letName, exprTpe),
-              body = inExpr,
-              tpe = inExpr.tpe
-            )
-      }
-    } yield r
-
-  private def compileRef(ctx: CompilerContext, ref: Expressions.REF): SetTypeResult[EXPR] = EitherT.fromEither {
-    ref.key.toEither.flatMap { key =>
-      ctx.varDefs
-        .get(key)
-        .map(REF(key, _))
-        .toRight(s"A definition of '$key' is not found")
-    }
-  }
-
-  private def compileMatch(ctx: CompilerContext, m: Expressions.MATCH): SetTypeResult[EXPR] = {
-    val Expressions.MATCH(_, _, expr, cases) = m
-    val rootMatchTmpArg                      = "$match" + ctx.tmpArgsIdx
-    val updatedCtx                           = ctx.copy(tmpArgsIdx = ctx.tmpArgsIdx + 1)
-    val typeDefs                             = ctx.predefTypes
-    def flat(tl: List[String]): List[String] = {
-      tl.flatMap { t =>
-        typeDefs.get(t).fold(List(t)) {
-          case UnionType(_, tl) => tl.map(_.name)
-          case t                => List(t.name)
-        }
-      }
-    }
-
-    for {
-      typedExpr <- compile(ctx, expr)
-      possibleExpressionTypes <- EitherT.fromEither[Coeval](typedExpr.tpe match {
-        case u: UNION => Right(u)
-        case _        => Left("Only union type can be matched")
+            .toCompileM
+        })
+      refTmpKey = "$match" + ctx.tmpArgsIdx
+      _ <- modify[CompilerContext, CompilationError](c => c.copy(tmpArgsIdx = c.tmpArgsIdx + 1))
+      ifCases <- inspect[CompilerContext, CompilationError, Expressions.EXPR](updatedCtx => {
+        mkIfCases(updatedCtx, cases, Expressions.REF(1, 1, PART.VALID(1, 1, refTmpKey)))
       })
-      matchingTypes <- EitherT.fromEither[Coeval](cases.flatMap(_.types).map(_.toEither).toList.sequence[CompilationResult, String])
-      matchedTypes = UNION(flat(matchingTypes).map(CASETYPEREF))
-      lastEmpty    = cases.last.types.isEmpty
-      _ <- EitherT.cond[Coeval](
-        lastEmpty && (possibleExpressionTypes >= matchedTypes) || (possibleExpressionTypes equivalent matchedTypes),
-        (),
-        s"Matching not exhaustive: possibleTypes are ${possibleExpressionTypes.l}, while matched are $matchingTypes"
-      )
-      refTmp = Expressions.REF(1, 1, PART.VALID(1, 1, rootMatchTmpArg))
-      ifBasedCases: Expressions.EXPR = cases.foldRight(Expressions.REF(1, 1, PART.VALID(1, 1, PureContext.errRef)): Expressions.EXPR) {
-        case (mc, further) =>
-          val blockWithNewVar = mc.newVarName match {
-            case Some(newVal) => Expressions.BLOCK(1, 1, Expressions.LET(1, 1, newVal, refTmp, mc.types), mc.expr)
-            case None         => mc.expr
-          }
-          mc.types.toList match {
-            case Nil => blockWithNewVar
-            case types =>
-              def isInst(matchType: String): Expressions.EXPR =
-                Expressions.FUNCTION_CALL(1,
-                                          1,
-                                          PART.VALID(1, 1, PureContext._isInstanceOf.name),
-                                          List(refTmp, Expressions.CONST_STRING(1, 1, PART.VALID(1, 1, matchType))))
-              val hType :: tTypes = flat(types.map(_.asInstanceOf[PART.VALID[String]].v))
-              val typeIf = tTypes.foldLeft(isInst(hType)) {
-                case (other, matchType) => BINARY_OP(1, 1, isInst(matchType), BinaryOperation.OR_OP, other)
-              }
-
-              val blockWithNewVar = mc.newVarName match {
-                case Some(newVal) => Expressions.BLOCK(1, 1, Expressions.LET(1, 1, newVal, refTmp, mc.types), mc.expr)
-                case None         => mc.expr
-              }
-              Expressions.IF(1, 1, typeIf, blockWithNewVar, further)
-          }
-      }
-      compiled <- compileBlock(updatedCtx,
-                               Expressions.BLOCK(1, 1, Expressions.LET(1, 1, PART.VALID(1, 1, rootMatchTmpArg), expr, Seq.empty), ifBasedCases))
-    } yield compiled
+      compiledMatch <- compileBlock(start, end, Expressions.LET(1, 1, PART.VALID(1, 1, refTmpKey), expr, Seq.empty), ifCases)
+    } yield compiledMatch
   }
 
-  private def resolvedFuncArguments(ctx: CompilerContext, args: List[Expressions.EXPR]): ResolvedArgsResult = {
-    import cats.instances.list._
-    val r: List[SetTypeResult[EXPR]] = args.map(arg => compile(ctx, arg))(collection.breakOut)
-    r.sequence[SetTypeResult, EXPR]
+  private def compileBlock(start: Int, end: Int, let: Expressions.LET, body: Expressions.EXPR): CompileM[Terms.EXPR] = {
+    for {
+      ctx <- get[CompilerContext, CompilationError]
+      letName <- handlePart(let.name)
+        .ensureOr(n => AlreadyDefined(start, end, n, false))(n => !ctx.varDefs.contains(n))
+        .ensureOr(n => AlreadyDefined(start, end, n, true))(n => !ctx.functionDefs.contains(n))
+      compiledLet <- compileExpr(let.value)
+      letTypes <- let.types.toList
+        .traverse[CompileM, String](handlePart)
+        .ensure(NonExistingType(start, end, letName, ctx.predefTypes.keys.toList))(_.forall(ctx.predefTypes.contains))
+      typeUnion = if (letTypes.isEmpty) compiledLet.tpe else UNION(letTypes.map(CASETYPEREF))
+      _            <- modify[CompilerContext, CompilationError](vars.modify(_)(_ + (letName -> typeUnion)))
+      compiledBody <- compileExpr(body)
+    } yield BLOCK(LET(letName, compiledLet), compiledBody, compiledBody.tpe)
   }
 
-  private def matchFuncOverload(funcName: String,
+  private def compileGetter(start: Int, end: Int, fieldPart: PART[String], refExpr: Expressions.EXPR): CompileM[Terms.EXPR] = {
+    for {
+      ctx         <- get[CompilerContext, CompilationError]
+      field       <- handlePart(fieldPart)
+      compiledRef <- compileExpr(refExpr)
+      result <- (compiledRef.tpe match {
+        case CASETYPEREF(name) => mkGetter(start, end, ctx, name, field, compiledRef)
+        case UNION(tl)         => mkGetter(start, end, ctx, tl, field, compiledRef)
+        case _                 => Generic(start, end, "Unexpected ref type: neither simple type nor union type").asLeft[EXPR]
+      }).toCompileM
+    } yield result
+  }
+
+  private def compileFunctionCall(start: Int, end: Int, namePart: PART[String], args: List[Expressions.EXPR]): CompileM[EXPR] = {
+    for {
+      name         <- handlePart(namePart)
+      signatures   <- get[CompilerContext, CompilationError].map(_.functionTypeSignaturesByName(name))
+      compiledArgs <- args.traverse(compileExpr)
+      result <- (signatures match {
+        case Nil           => FunctionNotFound(start, end, name, compiledArgs.map(_.tpe.toString)).asLeft[EXPR]
+        case single :: Nil => matchFuncOverload(start, end, name, args, compiledArgs, single)
+        case many => {
+          val matchedSigs = many
+            .map(matchFuncOverload(start, end, name, args, compiledArgs, _))
+            .collect({ case Right(ex) => ex })
+
+          matchedSigs match {
+            case Nil         => FunctionNotFound(start, end, name, compiledArgs.map(_.tpe.toString)).asLeft[EXPR]
+            case call :: Nil => call.asRight[CompilationError]
+            case _           => AmbiguousOverloading(start, end, name, signatures.toList).asLeft[EXPR]
+          }
+        }
+      }).toCompileM
+    } yield result
+  }
+
+  private def compileRef(start: Int, end: Int, keyPart: PART[String]): CompileM[EXPR] = {
+    for {
+      key <- handlePart(keyPart)
+      ctx <- get[CompilerContext, CompilationError]
+      result <- ctx.varDefs
+        .get(key)
+        .fold(raiseError[CompilerContext, CompilationError, EXPR](DefNotFound(start, end, key)))(t => (REF(key, t): EXPR).pure[CompileM])
+    } yield result
+  }
+
+  private def matchFuncOverload(start: Int,
+                                end: Int,
+                                funcName: String,
                                 funcArgs: List[Expressions.EXPR],
                                 resolvedArgs: List[EXPR],
-                                f: FunctionTypeSignature): Either[String, EXPR] = {
-    val argTypes   = f.args
-    val resultType = f.result
+                                f: FunctionTypeSignature): Either[CompilationError, EXPR] = {
+    val argTypes = f.args
     if (funcArgs.lengthCompare(argTypes.size) != 0)
-      Left(s"Function '$funcName' requires ${argTypes.size} arguments, but ${funcArgs.size} are provided")
+      Left(WrongArgumentsNumber(start, end, funcName, argTypes.size, funcArgs.size))
     else {
       val typedExpressionArgumentsAndTypedPlaceholders: List[(EXPR, TYPEPLACEHOLDER)] = resolvedArgs.zip(argTypes)
 
       val typePairs = typedExpressionArgumentsAndTypedPlaceholders.map { case (typedExpr, tph) => (typedExpr.tpe, tph) }
       for {
-        resolvedTypeParams <- TypeInferrer(typePairs)
-        resolvedResultType <- TypeInferrer.inferResultType(resultType, resolvedTypeParams)
-      } yield
-        FUNCTION_CALL(
-          FunctionHeader(funcName, f.args.map(FunctionHeaderType.fromTypePlaceholder)),
-          typedExpressionArgumentsAndTypedPlaceholders.map(_._1),
-          resolvedResultType
-        )
+        resolvedTypeParams <- TypeInferrer(typePairs).leftMap(Generic(start, end, _))
+        resolvedResultType <- TypeInferrer.inferResultType(f.result, resolvedTypeParams).leftMap(Generic(start, end, _))
+        header = FunctionHeader(funcName, f.args.map(FunctionHeaderType.fromTypePlaceholder))
+        args   = typedExpressionArgumentsAndTypedPlaceholders.map(_._1)
+      } yield FUNCTION_CALL(header, args, resolvedResultType): EXPR
     }
   }
 
-  private def handlePart[T](part: PART[T])(f: T => EXPR): SetTypeResult[EXPR] = part match {
-    case PART.VALID(_, _, x)               => EitherT.pure(f(x))
-    case PART.INVALID(start, end, message) => EitherT.leftT[Coeval, EXPR](s"$message in $start-$end")
+  def resolveFields(ctx: CompilerContext, tpe: PredefBase): List[(String, TYPE)] = tpe match {
+    case PredefCaseType(_, fields) => fields
+    case UnionType(_, types)       => types.map(n => resolveFields(ctx, ctx.predefTypes(n.name)).toSet).reduce(_ intersect _).toList
   }
 
-  def apply(c: CompilerContext, expr: Expressions.EXPR): CompilationResult[EXPR] = {
-    def result = compile(c, expr).value().left.map { e =>
-      s"Typecheck failed: $e"
-    }
-    Try(result) match {
-      case scala.util.Failure(ex)  => Left(ex.toString)
-      case scala.util.Success(res) => res
-    }
+  def mkIf(start: Int, end: Int, cond: EXPR, ifTrue: EXPR, ifFalse: EXPR): Either[CompilationError, EXPR] = {
+    if (cond.tpe != BOOLEAN) Left(UnexpectedType(start, end, "BOOLEAN", cond.tpe.toString))
+    else
+      TypeInferrer
+        .findCommonType(ifTrue.tpe, ifFalse.tpe)
+        .fold(UnexpectedType(start, end, ifTrue.tpe.toString, ifFalse.tpe.toString).asLeft[IF])(IF(cond, ifTrue, ifFalse, _).asRight)
+  }
+
+  def mkIfCases(ctx: CompilerContext, cases: List[MATCH_CASE], refTmp: Expressions.REF): Expressions.EXPR = {
+    cases.foldRight(Expressions.REF(1, 1, PART.VALID(1, 1, PureContext.errRef)): Expressions.EXPR)((mc, further) => {
+      val blockWithNewVar = mc.newVarName.fold(mc.expr)(nv => Expressions.BLOCK(1, 1, Expressions.LET(1, 1, nv, refTmp, mc.types), mc.expr))
+      mc.types.toList match {
+        case Nil => blockWithNewVar
+        case types =>
+          def isInst(matchType: String): Expressions.EXPR =
+            Expressions
+              .FUNCTION_CALL(1,
+                             1,
+                             PART.VALID(1, 1, PureContext._isInstanceOf.name),
+                             List(refTmp, Expressions.CONST_STRING(1, 1, PART.VALID(1, 1, matchType))))
+          val hType :: tTypes = flat(ctx.predefTypes, types.map(_.asInstanceOf[PART.VALID[String]].v))
+          val typeIf          = tTypes.foldLeft(isInst(hType))((other, matchType) => BINARY_OP(1, 1, isInst(matchType), BinaryOperation.OR_OP, other))
+          Expressions.IF(1, 1, typeIf, blockWithNewVar, further)
+      }
+    })
+  }
+
+  private def mkGetter(start: Int,
+                       end: Int,
+                       ctx: CompilerContext,
+                       typeName: String,
+                       fieldName: String,
+                       expr: EXPR): Either[CompilationError, GETTER] = {
+    for {
+      refTpe <- ctx.predefTypes
+        .get(typeName)
+        .fold(TypeNotFound(start, end, typeName).asLeft[PredefBase])(_.asRight)
+      fieldTpe <- resolveFields(ctx, refTpe)
+        .collectFirst({ case (field, tpe) if fieldName == field => tpe })
+        .fold(FieldNotFound(start, end, fieldName, typeName).asLeft[TYPE])(_.asRight)
+    } yield GETTER(expr, fieldName, fieldTpe)
+  }
+
+  private def mkGetter(start: Int,
+                       end: Int,
+                       ctx: CompilerContext,
+                       types: List[CASETYPEREF],
+                       fieldName: String,
+                       expr: EXPR): Either[CompilationError, GETTER] = {
+    types
+      .traverse[Option, Terms.TYPE](ctr => {
+        ctx.predefTypes
+          .get(ctr.name)
+          .flatMap(resolveFields(ctx, _).find(_._1 == fieldName).map(_._2))
+      })
+      .fold((FieldNotFound(start, end, fieldName, s"Union($types)"): CompilationError).asLeft[GETTER])(ts => {
+        TypeInferrer.findCommonType(ts) match {
+          case Some(ct) => GETTER(expr, fieldName, ct).asRight[CompilationError]
+          case None     => UnexpectedType(start, end, s"UNION($types)", s"UNION($ts)").asLeft[GETTER]
+        }
+      })
+  }
+
+  private def handlePart[T](part: PART[T]): CompileM[T] = part match {
+    case PART.VALID(_, _, x)               => x.pure[CompileM]
+    case PART.INVALID(start, end, message) => raiseError(Generic(start, end, message))
+  }
+  def apply(c: CompilerContext, expr: Expressions.EXPR): Either[String, EXPR] = {
+    compileExpr(expr)
+      .run(c)
+      .map(_._2.leftMap(e => s"Typecheck failed: ${Show[CompilationError].show(e)}"))
+      .value
   }
 }
