+++ lang/shared/src/main/scala/com/wavesplatform/lang/v1/compiler/package.scala
@@ -1,11 +1,11 @@
 package com.wavesplatform.lang.v1
 
-import com.wavesplatform.lang.v1.task.{TaskM, TaskMFunctions}
 import cats.implicits._
-import task.imports._
+import com.wavesplatform.lang.v1.task.TaskM
+import com.wavesplatform.lang.v1.task.imports._
 
 package object compiler {
-  type CompileM[A]      = TaskM[CompilerContext, CompilationError, A]
+  type CompileM[A] = TaskM[CompilerContext, CompilationError, A]
 
   implicit class EiExt[A](ei: Either[CompilationError, A]) {
     def toCompileM: CompileM[A] =
